drivers: qcom: smp2p: add a timer to the held wakelock
On adding logs to the wakelock addition/removal code in the wakeup driver, it can
be found that the smp2p wakelock is held for 2ms at most. This wakelock
for some weird reason gets stuck at times and doesn't let the device
deep sleep. This commit adds a timeout of 250ms to
the wakelock, which is already far more than what it should take to
finish it's job.