thermal: cpu_cooling: Add cmdline for disabling mi thermal helper
When userspace is not using mi thermald (e.g. pixel thermal hal),
platform based cpu mitigation is expected to be totally usable, but
it's not the case anymore when mi thermal helper is exists since it
just straightly override the platform current state with the value from
mi thermald.

This adds cmdline to disable mi thermal helper and bring back things
that changed by it.